BREAKING: interface config moved under interfaces
Previously, some interface config was defined under the devices dictionary in the yaml file.  We have moved all interface configuration under the interfaces dictionary, and changed the interface value in the devices dictionary to be a pointer to the interface configuration.  This broke the device and ip_address libraries and related scripts, so we've updated these as well.

We also changed mgmt_ip, which was under devices, to ip4, which is now under interfaces.  We've changed relevant scripts args to match as well.