run through black and pylint, more...
Retrieve valid choices from user's netbox instance.
Support role parameter.
Remove unused imports.

For create and update operations, this class now also maps and assigns the IP address as a device's primary IP.  We shouldn't have to do this, but are seeing Netbox get into a state where the IPAM -> IP Addresses page will not load if the IP address is not mapped to a device.