Support ip address options, more...
We've made a small change to the YAML, described below.

Previously, the interfaces->ip4 field in example.yml was used solely as an ip4 address.  Now, it is also used as a pointer into a new dictionary;  ip4_addresses.  If the value of interfaces->ip4 is found as a key in ip4_addresses, then the entry associated with this key is used to configure other ip4 parameters (currently, role and status).  If the ip4 value is NOT found in ip4_addresses dictionary, then we revert to the  previous usage (i.e. that of an ip4 address).